refactor: Modernize and simplify Svelte components
